pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "terraria-agent"
version = "0.1.0"
description = "Pure vision-driven Terraria AI agent"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyautogui>=0.9",
]

[project.optional-dependencies]
vision = [
    "mss>=9.0",
    "numpy>=1.26",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "pyobjc-framework-Quartz>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/terraria_agent"]
